+++ packages/terra-navigation-prompt/src/terra-dev-site/doc/navigation-prompt/examples/NavigationPromptExample.jsx
@@ -2,7 +2,7 @@ import React, { useState, useRef } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import NavigationPrompt, { NavigationPromptCheckpoint } from '../../../../index';
-import styles from './NavigationPromptExample.scss';
+import styles from './NavigationPromptExample.module.scss';
 
 const cx = classNames.bind(styles);