+++ packages/terra-popup/src/terra-dev-site/test/popup/ArrowHorizontalAttachmentsPopup.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 import Popup from '../../../Popup';
-import styles from './ArrowHorizontalAttachmentsPopup.test.scss';
+import styles from './ArrowHorizontalAttachmentsPopup.test.module.scss';
 
 const cx = classNames.bind(styles);