+++ packages/terra-navigation-layout/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Removed all inline styles from Dev-site components and implemented the same styles using external css.
+* Renamed test and example SCSS files to end with `.module.scss` extension
 
 5.4.0 - (June 19, 2019)
 ------------------